refactored the logic that applies color customizations and that writes them to settings, to apply-color.ts. This is the heart of the Peacock code and hopefully makes it easier to follow for future refactoring.
